updated sidebar and scrolling
Import MouseEventHandler
